requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astunparse==1.6.3
beautifulsoup4==4.9.3
bootstrap4==0.1.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-bootstrap4==3.0.1
django-heroku==0.3.1
flatbuffers==1.12
future==0.18.2
gast==0.2.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
Markdown==3.3.4
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
style==1.1.0
#tensorflow==2.4.1
tensorflow==1.15.5
tensorflow-estimator==1.15.0
termcolor==1.1.0
tflite==2.1.0
typing-extensions==3.7.4.3
update==0.0.1
urllib3==1.26.4
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
django-storages==1.11.1
boto3==1.17.64
psycopg2-binary
